server/requirements.txt
Pillow>=10.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1
cloudinary>=1.35.0
fastapi>=0.104.0
flask-cors==3.0.10
flask==2.2.3
folium==0.14.0
jinja2>=3.1.2
joblib==1.2.0
joblib>=1.3.1
matplotlib==3.7.1
matplotlib>=3.7.2
numpy==1.24.2
numpy>=1.24.3
pandas==1.5.3
pandas>=2.0.3
pillow==9.5.0
pydantic>=2.4.2
pymongo>=4.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests==2.28.2
requests>=2.31.0
scikit-learn==1.2.2
scikit-learn>=1.3.0
starlette>=0.31.1
streamlit-folium==0.11.1
streamlit==1.20.0
tensorflow==2.13.0
tensorflow>=2.15.0
uvicorn>=0.23.2
